refactor upload functions and ShowRenderImage operator
