+++ __init__.py
@@ -48,6 +48,7 @@ from .src import async_loop, autothumb
 from .src.application.application import Application
 from .src.authentication.authentication import Authentication
 from .src.ui import render as ui_render
+from .src.ui.operators import render_image
 
 bl_info = {
     'name': 'Hana3D',
@@ -310,6 +311,7 @@ modules = (
     libraries,
     logger,
     render,
+    render_image,
     search,
     tags,
     tasks_queue,
