+++ src/ui/callbacks/asset_bar.py
@@ -5,11 +5,12 @@ import os
 
 import bpy
 
-from .. import bgl_helper
-from ...preferences.preferences import Preferences
-from ...search.search import Search
 from .... import paths, utils
 from ....config import HANA3D_NAME, HANA3D_UI
+from ...preferences.preferences import Preferences
+from ...search.search import Search
+from ...upload import upload
+from .. import bgl_helper
 
 verification_icons = {
     'ready': 'vs_ready.png',
@@ -180,7 +181,7 @@ def draw_tooltip(x, y, text='', author='', img=None, gravatar=None):
 
 def draw_callback_2d_upload_preview(self, context):
     ui_props = getattr(context.window_manager, HANA3D_UI)
-    props = utils.get_upload_props()
+    props = upload.get_upload_props()
 
     if props is not None and ui_props.draw_tooltip:
         ui_props.thumbnail_image = props.thumbnail
