+++ src/ui/operators/__init__.py
@@ -3,6 +3,7 @@
 
 from .asset_bar import AssetBarOperator
 from .default_names import DefaultNamesOperator
+from .render_image import ShowRenderImage
 from .run_asset_bar import RunAssetBarWithContext
 from .transfer_data import TransferHana3DData
 from .undo import UndoWithContext
