+++ thread_tools.py
@@ -66,8 +66,8 @@ def update_renders_in_foreground(asset_type: str, view_id: str):
         asset_type: str (model|scene|material)
         view_id: str
     """
-    imports = 'from . import utils, render_tools'
-    props = f'props = utils.get_upload_props_by_view_id("{asset_type}", "{view_id}")'
+    imports = 'from . import render_tools; from .src.upload import upload'
+    props = f'props = upload.get_upload_props_by_view_id("{asset_type}", "{view_id}")'
     update = 'render_tools.update_render_list(props, set_jobs=False)'
     cmd = f'{imports}; {props}; {update}'
     state_update_queue.put(cmd)
